brush creates from  bottom up as to not have the object spaz out
frame rate is now only shown as an int

+++ Code Test/Code/Render.h
@@ -18,7 +18,7 @@ void Render()
 	brushSize.Draw(mainRenderer, 20, 54);
 
 	//render the current framerate
-	currentFrameRate.SetText(std::to_string(avgFPS));
+	currentFrameRate.SetText(std::to_string(static_cast<int>(avgFPS)));
 	currentFrameRate.Draw(mainRenderer, 20, WINDOW_HEIGHT - 34);
 
 	SDL_RenderPresent(mainRenderer);
